+++ mall-admin/pom.xml
@@ -29,6 +29,10 @@
             <artifactId>mall-mbg</artifactId>
         </dependency>
         <dependency>
+            <groupId>com.macro.mall</groupId>
+            <artifactId>mall-security</artifactId>
+        </dependency>
+        <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-security</artifactId>
         </dependency>
